+++ api/test/test-utils.js
@@ -10,7 +10,10 @@ const datasetorderModel = require('../src/models/datasetorderModel');
 const workerpoolorderModel = require('../src/models/workerpoolorderModel');
 const requestorderModel = require('../src/models/requestorderModel');
 
-const sleep = (ms) => new Promise((res) => setTimeout(res, ms));
+const sleep = (ms) =>
+  new Promise((res) => {
+    setTimeout(res, ms);
+  });
 
 const WALLETS = {
   DEFAULT: {
@@ -65,7 +68,7 @@ const deployAppFor = async (iexec, owner) => {
     multiaddr: 'registry.hub.docker.com/iexechub/vanityeth:1.1.1',
     checksum:
       '0x00f51494d7a42a3c1c43464d9f09e06b2a99968e3b978f6cd11ab3410b7bcd14',
-    mrenclave: 'abc|123|test',
+    mrenclave: '',
   });
   return address;
 };